web/backend_flask/requirements.txt
Flask
paho-mqtt
python-dotenv

